backend/requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==5.5.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
curl_cffi==0.15.0
cycler==0.12.1
fastapi==0.135.3
filelock==3.20.2
finnhub-python==2.4.27
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
frozendict==2.4.7
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
importlib_metadata==7.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
multitasking==0.0.12
narwhals==2.17.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
peewee==4.0.4
pillow==10.4.0
platformdirs==4.9.4
plotly==5.18.0
protobuf==4.25.8
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
scikit-learn==1.4.0
scipy==1.16.3
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
starlette==1.0.0
streamlit==1.31.0
tenacity==8.5.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.44.0
validators==0.35.0
websockets==16.0
Werkzeug==3.1.4
xgboost==3.2.0
xlsxwriter==3.2.9
yfinance==1.2.1
zipp==3.23.0
